+++ core/src/main/java/de/raysha/lib/jsimpleshell/CommandRecorder.java
@@ -8,6 +8,7 @@ import de.raysha.lib.jsimpleshell.exception.ExitException;
 import de.raysha.lib.jsimpleshell.handler.CommandAccessManager;
 import de.raysha.lib.jsimpleshell.handler.CommandAccessManager.AccessDecision.Decision;
 import de.raysha.lib.jsimpleshell.handler.CommandLoopObserver;
+import de.raysha.lib.jsimpleshell.script.Variable;
 
 /**
  * This class is responsible for recording command lines without execute them.
@@ -15,6 +16,8 @@ import de.raysha.lib.jsimpleshell.handler.CommandLoopObserver;
  * @author rainu
  */
 public class CommandRecorder implements CommandAccessManager, CommandLoopObserver {
+	public static final String VARIABLE_NAME_RECORD_MODE = "_RECORD_MODE";
+
 	private final Shell shell;
 	private final List<String> commands = new LinkedList<String>();
 	private boolean inCommandLoop = false;
@@ -35,12 +38,38 @@ public class CommandRecorder implements CommandAccessManager, CommandLoopObserve
 		inCommandLoop = true;
 
 		try{
+			inserRecordModeVariable();
 			shell.commandLoop();
 		}finally{
+			removeRecordModeVariable();
 			inCommandLoop = false;
 		}
 
 		return commands;
+	}
+
+	private void inserRecordModeVariable() {
+		shell.getEnvironment().setVariable(new Variable(VARIABLE_NAME_RECORD_MODE, true, true));
+	}
+
+	private void removeRecordModeVariable() {
+		shell.getEnvironment().setVariable(new Variable(VARIABLE_NAME_RECORD_MODE, false, true));
+	}
+
+	/**
+	 * Checks whether the given shell is in record mode.
+	 * In record mode no persistent actions should be done!
+	 *
+	 * @param shell The shell to check.
+	 * @return True if the shell is in record mode. Otherwise false.
+	 */
+	public static boolean isShellInRecordMode(Shell shell){
+		Variable var = shell.getEnvironment().getVariable(VARIABLE_NAME_RECORD_MODE);
+		if(var == null){
+			return false;
+		}
+
+		return (Boolean)var.getValue();
 	}
 
 	@Override